arch.x86-64.intr: fix `cpu_idx` incorrect type: `u8` to `u16`

+++ src/kernel/arch/x86-64/gdt.zig
@@ -119,7 +119,7 @@ pub fn addTss(tss: *const intr.TaskStateSegment) !void {
     );
 }
 
-pub inline fn getTssOffset(idx: u8) u16 {
+pub inline fn getTssOffset(idx: u16) u16 {
     return @truncate((tss_base_idx * @sizeOf(SegmentDescriptor)) + (idx * @sizeOf(SystemSegmentDescriptor)));
 }
